refactor(particle-system): extract types and constants to separate files and update imports (#2052)
* refactor(particle-system): separate types from component for better maintainability

* refactor(particle-system): add ir description to types

* refactor(particle-system): separate constants from component for better maintainability

* refactor(particle-system): add ir description to const files

* refactor(particle-system): update imports to use extracted types and constants

* refactor(particle-system): remove constants and type files

* refactor(particle-system): move types, const and default to the type file

* refactor(particle-system): update imports
